StatusProxy: Factor out watchdog update timer.
